Split visual polish Playwright spec
